+++ backend/app/bindings/api/v1/bindings/stage.rb
@@ -51,7 +51,7 @@ module Api::V1::Bindings
         :'order' => :'Integer',
         :'title' => :'String',
         :'description' => :'String',
-        :'available_after_days' => :'Integer',
+        :'available_after_days' => :'Float',
         :'config' => :'Object'
       }
     end
